// hr-frontend/src/lib/api/generated/model/functionalRelationTypeSetting.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface FunctionalRelationTypeSetting {
  functionalRelationTypeId?: number;
  code?: string;
  nameEn?: string;
  nameAr?: string;
  description?: string;
  approvalRelation?: boolean;
  active?: boolean;
  usageCount?: number;
  createdAt?: string;
  updatedAt?: string;
}